Moved checkpoint utility in its own directory
